+++ src/main/java/jpa/ParticipantRepository.java
@@ -5,6 +5,6 @@ import jpa.Participant;
 
 // This will be AUTO IMPLEMENTED by Spring into a Bean
 // CRUD refers Create, Read, Update, Delete
-public interface ParticipantRepository extends CrudRepository<Participant, Integer> {
+public interface ParticipantRepository extends CrudRepository<Participant, Long> {
 
 }
